feat: verifyページの文字列を定義
