+++ web/user/src/locales/tsconfig.json
@@ -4,7 +4,7 @@
     "module": "esnext",
     "moduleResolution": "node",
     "sourceMap": false,
-    "rootDir": ".",
+    "rootDir": "..",
     "outDir": ".",
     "skipLibCheck": true
   },
