+++ web/user/src/types/locales/i18n.ts
@@ -1,0 +1,16 @@
+export interface I18n {
+  auth: {
+    signUp: {
+      tel: string
+      email: string
+      password: string
+      passwordConfirm: string
+      signUp: string
+      alreadyHas: string
+    }
+    verify: {
+      message: string
+      btnText: string
+    }
+  }
+}
