+++ web/user/src/types/locales/index.ts
@@ -1,12 +1,1 @@
-export interface I18n {
-  auth: {
-    signUp: {
-      tel: string
-      email: string
-      password: string
-      passwordConfirm: string
-      signUp: string
-      alreadyHas: string
-    }
-  }
-}
+export * from './i18n'
